% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdisplot.R
\name{ggdisplot}
\alias{ggdisplot}
\title{Plot scatterplot of two variables along with two density plots.}
\usage{
ggdisplot(df, var1, var2, color, xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{df}{dataframe to work with.}

\item{var1}{x variable to plot (string).}

\item{var2}{y variable to plot (string).}

\item{color}{factor variable for ggplot (string).}

\item{xlim}{limits for x variable eg: c(1, 10).}

\item{ylim}{limits for y variable eg: c(1, 10).}

\item{xlab}{label for x variable.}

\item{ylab}{label for y variable.}

\item{...}{additional ggplot parameters.}
}
\value{
plot as returned by cowplot::plot_grid()
}
\description{
Plot scatterplot of two variables along with two density plots.
}
\examples{
ggdisplot(iris, "Sepal.Length", "Sepal.Width", "Species")
ggdisplot(iris, "Sepal.Length", "Sepal.Width", "Species", xlim=c(1,10), ylim=c(1,10))
ggdisplot(iris, "Sepal.Length", "Sepal.Width", "Species", theme_bw())
}
